Add TSL unittest
IB-4631, IB-4621
