+++ test/CMakeLists.txt
@@ -10,22 +10,61 @@ if( Boost_INCLUDE_DIR )
         ${XERCESC_INCLUDE_DIR}
         ${Boost_INCLUDE_DIR}
     )
-    if( LIBDIGIDOC_FOUND AND LIBDIGIDOC_LINKED )
-        list( APPEND unittests_DEFS LINKED_LIBDIGIDOC )
-    endif()
-    add_executable( unittests libdigidocpp_boost.cpp )
-    set_target_properties( unittests PROPERTIES
-        COMPILE_DEFINITIONS "DIGIDOCPPCONF=\"${CMAKE_SOURCE_DIR}/etc/schema\";PKCS11_MODULE=\"${PKCS11_MODULE}\";${unittests_DEFS}"
-        LINK_LIBRARIES "digidocpp;digidocpp_priv"
+    add_definitions(
+        -DDIGIDOCPPCONF="${CMAKE_SOURCE_DIR}/etc/schema"
+        -DPKCS11_MODULE="${PKCS11_MODULE}"
     )
+    if( LIBDIGIDOC_FOUND AND LIBDIGIDOC_LINKED )
+        add_definitions(LINKED_LIBDIGIDOC)
+    endif()
+    add_executable(unittests libdigidocpp_boost.cpp)
+    add_executable(TSLTests TSLTests.cpp)
+    target_link_libraries(unittests digidocpp digidocpp_priv)
+    target_link_libraries(TSLTests digidocpp)
     if(WIN32)
         string(REPLACE "/EHsc" "/EHa" CMAKE_CXX_FLAGS ${CMAKE_CXX_FLAGS})
     else()
-        set_target_properties( unittests PROPERTIES COMPILE_FLAGS "-Wno-unused-parameter -Wno-unused-variable" )
+        set_target_properties(unittests TSLTests PROPERTIES COMPILE_FLAGS "-Wno-unused-parameter -Wno-unused-variable")
     endif()
     add_test(NAME runtest
         COMMAND ${CMAKE_CURRENT_BINARY_DIR}/unittests --build_info=YES --output_format=XML --log_level=all --report_level=no --log_sink=${CMAKE_CURRENT_BINARY_DIR}/libdigidocpp.xml -- ${CMAKE_CURRENT_SOURCE_DIR}/data
         WORKING_DIRECTORY ${CMAKE_BINARY_DIR}/src
     )
-    add_custom_target(check ${RUN_TESTS} COMMAND ${CMAKE_CTEST_COMMAND} DEPENDS unittests)
+    add_test(NAME TSLTest_CA-withdrawn
+        COMMAND ${CMAKE_CURRENT_BINARY_DIR}/TSLTests -- EE_T-CA-withdrawn.xml bad ${CMAKE_CURRENT_SOURCE_DIR}/data
+        WORKING_DIRECTORY ${CMAKE_BINARY_DIR}/src
+    )
+    add_test(NAME TSLTest_CA-withdrawn-granted-before
+        COMMAND ${CMAKE_CURRENT_BINARY_DIR}/TSLTests -- EE_T-CA-withdrawn-granted-before.xml good ${CMAKE_CURRENT_SOURCE_DIR}/data
+        WORKING_DIRECTORY ${CMAKE_BINARY_DIR}/src
+    )
+    add_test(NAME TSLTest_CA-withdrawn-granted-after
+        COMMAND ${CMAKE_CURRENT_BINARY_DIR}/TSLTests -- EE_T-CA-withdrawn-granted-after.xml bad ${CMAKE_CURRENT_SOURCE_DIR}/data
+        WORKING_DIRECTORY ${CMAKE_BINARY_DIR}/src
+    )
+    add_test(NAME TSLTest_OCSP-withdrawn
+        COMMAND ${CMAKE_CURRENT_BINARY_DIR}/TSLTests -- EE_T-OCSP-withdrawn.xml bad ${CMAKE_CURRENT_SOURCE_DIR}/data
+        WORKING_DIRECTORY ${CMAKE_BINARY_DIR}/src
+    )
+    add_test(NAME TSLTest_OCSP-withdrawn-granted-before
+        COMMAND ${CMAKE_CURRENT_BINARY_DIR}/TSLTests -- EE_T-OCSP-withdrawn-granted-before.xml good ${CMAKE_CURRENT_SOURCE_DIR}/data
+        WORKING_DIRECTORY ${CMAKE_BINARY_DIR}/src
+    )
+    add_test(NAME TSLTest_OCSP-withdrawn-granted-after
+        COMMAND ${CMAKE_CURRENT_BINARY_DIR}/TSLTests -- EE_T-OCSP-withdrawn-granted-after.xml bad ${CMAKE_CURRENT_SOURCE_DIR}/data
+        WORKING_DIRECTORY ${CMAKE_BINARY_DIR}/src
+    )
+    add_test(NAME TSLTest_TSA-withdrawn
+        COMMAND ${CMAKE_CURRENT_BINARY_DIR}/TSLTests -- EE_T-TSA-withdrawn.xml bad ${CMAKE_CURRENT_SOURCE_DIR}/data
+        WORKING_DIRECTORY ${CMAKE_BINARY_DIR}/src
+    )
+    add_test(NAME TSLTest_TSA-withdrawn-granted-before
+        COMMAND ${CMAKE_CURRENT_BINARY_DIR}/TSLTests -- EE_T-TSA-withdrawn-granted-before.xml good ${CMAKE_CURRENT_SOURCE_DIR}/data
+        WORKING_DIRECTORY ${CMAKE_BINARY_DIR}/src
+    )
+    add_test(NAME TSLTest_TSA-withdrawn-granted-after
+        COMMAND ${CMAKE_CURRENT_BINARY_DIR}/TSLTests -- EE_T-TSA-withdrawn-granted-after.xml bad ${CMAKE_CURRENT_SOURCE_DIR}/data
+        WORKING_DIRECTORY ${CMAKE_BINARY_DIR}/src
+    )
+    add_custom_target(check ${RUN_TESTS} COMMAND ${CMAKE_CTEST_COMMAND} DEPENDS unittests TSLTests)
 endif()
