+++ test/data/README.md
@@ -1,0 +1,16 @@
+# Test data
+
+## Container tests
+
+## TSL tests
+Validates [tsl.asice](tsl.asice) (signing time 2016-11-28T13:46:41Z) file with given TSL-s.
+
+* [EE_T-CA-withdrawn.xml](EE_T-CA-withdrawn.xml) - CA service revoked at 2016-11-27T21:00:00Z
+* [EE_T-CA-withdrawn-granted-before.xml](EE_T-CA-withdrawn-granted-before.xml) - CA service revoked at 2016-11-27T22:00:00Z and granted at 2016-11-28T01:00:00Z
+* [EE_T-CA-withdrawn-granted-later.xml](EE_T-CA-withdrawn-granted-later.xml) - CA service revoked at 2016-11-27T22:00:00Z and granted at 2016-11-29T01:00:00Z
+* [EE_T-OCSP-withdrawn.xml](EE_T-OCSP-withdrawn.xml) - OCSP service revoked at 2016-11-27T22:00:00Z
+* [EE_T-OCSP-withdrawn-granted-before.xml](EE_T-OCSP-withdrawn-granted-before.xml) - OCSP service revoked at 2016-11-27T22:00:00Z and granted at 2016-11-28T01:00:00Z
+* [EE_T-OCSP-withdrawn-granted-later.xml](EE_T-OCSP-withdrawn-granted-later.xml) - OCSP service revoked at 2016-11-27T22:00:00Z and granted at 2016-11-29T01:00:00Z
+* [EE_T-TSA-withdrawn.xml](EE_T-TSA-withdrawn.xml) - TSA service revoked at 2016-11-27T21:00:00Z
+* [EE_T-TSA-withdrawn-granted-before.xml](EE_T-TSA-withdrawn-granted-before.xml) - TSA service revoked at 2016-05-31T21:00:00Z and granted at 2016-11-28T01:00:00Z
+* [EE_T-TSA-withdrawn-granted-later.xml](EE_T-TSA-withdrawn-granted-later.xml) - TSA service revoked at 2014-11-27T21:00:00Z and granted at 2016-11-29T01:00:00Z
